move replyMessage to containers as Reply
